refactor: move AppStart/AppStartResult into domain/AppStart.kt
AppStartUtils.kt was left holding only these two value types after
Task 1 extracted checkAppStart into CheckAppStartUseCase, so the
"Utils" name no longer matched its content (every other *Utils.kt in
this codebase means "top-level extension functions"). Relocating them
next to their sole use case's package also fixes their layering: they
are domain-layer concepts that were sitting in the top-level package.

Moved AppStartTest.kt into the domain/ test package to match.
